# Add AsciiUtility, with workhorse methods for ASCII transcoding