This PR introduces internal workhorse methods for operating on buffers containing ASCII data (as bytes or chars). In particular, these methods provide implementations of:

- Counting the number of ASCII bytes or chars at the beginning of a sequence; and
- Transcoding ASCII data to and from UTF-16.

These workhorse methods will be used internally by the UTF-8 transcoder. However, there is opportunity for other components in the Framework to take direct dependencies on them. For example, I have a separate experimental branch that replumbs the existing `ASCIIEncoding` class to ride on top of these workhorse methods rather than its own existing implementation. Performance numbers from that branch are provided below. (Due to time constraints I only ran the numbers for UTF-16 to ASCII, not the other way around.)

|                Method | Toolchain | StringLength |        Mean |      Error |     StdDev |      Median | Ratio | RatioSD |
|---------------------- |---------- |------------- |------------:|-----------:|-----------:|------------:|------:|--------:|
|              **GetBytes** |     **local** |            **0** |    **791.6 ns** |   **6.483 ns** |   **5.414 ns** |    **792.5 ns** |  **0.83** |    **0.01** |
|              GetBytes |    master |            0 |    954.4 ns |   8.130 ns |   7.207 ns |    953.8 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            0 |    316.0 ns |   2.578 ns |   2.286 ns |    316.6 ns |  0.54 |    0.01 |
|          GetByteCount |    master |            0 |    583.8 ns |   5.517 ns |   4.607 ns |    584.9 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |            **1** |    **818.9 ns** |   **8.007 ns** |   **7.490 ns** |    **816.7 ns** |  **0.81** |    **0.01** |
|              GetBytes |    master |            1 |  1,013.8 ns |  10.640 ns |   9.952 ns |  1,012.4 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            1 |    424.1 ns |   5.297 ns |   4.955 ns |    424.5 ns |  0.73 |    0.01 |
|          GetByteCount |    master |            1 |    578.7 ns |   4.226 ns |   3.953 ns |    578.1 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |            **2** |    **844.1 ns** |   **7.613 ns** |   **7.122 ns** |    **842.6 ns** |  **0.72** |    **0.02** |
|              GetBytes |    master |            2 |  1,170.7 ns |  22.793 ns |  31.200 ns |  1,160.9 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            2 |    417.8 ns |   5.661 ns |   5.019 ns |    417.8 ns |  0.74 |    0.01 |
|          GetByteCount |    master |            2 |    564.2 ns |   4.996 ns |   4.674 ns |    564.3 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |            **3** |    **836.1 ns** |  **10.341 ns** |   **9.673 ns** |    **834.7 ns** |  **0.72** |    **0.01** |
|              GetBytes |    master |            3 |  1,161.3 ns |   5.638 ns |   4.998 ns |  1,160.9 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            3 |    409.1 ns |   4.238 ns |   3.757 ns |    408.8 ns |  0.73 |    0.01 |
|          GetByteCount |    master |            3 |    558.2 ns |   8.374 ns |   7.833 ns |    557.8 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |            **4** |    **825.7 ns** |  **11.601 ns** |  **10.852 ns** |    **824.4 ns** |  **0.64** |    **0.01** |
|              GetBytes |    master |            4 |  1,300.4 ns |  15.532 ns |  14.528 ns |  1,294.5 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            4 |    421.3 ns |   4.119 ns |   3.853 ns |    420.2 ns |  0.72 |    0.01 |
|          GetByteCount |    master |            4 |    588.4 ns |  13.071 ns |  12.838 ns |    584.2 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |            **8** |    **917.1 ns** |   **8.041 ns** |   **7.522 ns** |    **917.0 ns** |  **0.55** |    **0.01** |
|              GetBytes |    master |            8 |  1,679.6 ns |  12.461 ns |  11.656 ns |  1,679.4 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |            8 |    420.7 ns |   4.198 ns |   3.721 ns |    420.6 ns |  0.72 |    0.01 |
|          GetByteCount |    master |            8 |    581.1 ns |   3.444 ns |   3.222 ns |    581.3 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |           **12** |    **985.1 ns** |  **10.078 ns** |   **9.427 ns** |    **983.3 ns** |  **0.45** |    **0.01** |
|              GetBytes |    master |           12 |  2,199.9 ns |  43.483 ns |  40.674 ns |  2,200.9 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |           12 |    419.2 ns |   3.721 ns |   3.299 ns |    418.7 ns |  0.72 |    0.01 |
|          GetByteCount |    master |           12 |    581.4 ns |   7.347 ns |   6.513 ns |    582.8 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |           **16** |  **1,079.9 ns** |  **15.328 ns** |  **13.588 ns** |  **1,080.8 ns** |  **0.31** |    **0.04** |
|              GetBytes |    master |           16 |  3,609.7 ns | 134.463 ns | 396.467 ns |  3,626.6 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |           16 |    421.0 ns |   4.305 ns |   3.816 ns |    422.9 ns |  0.73 |    0.01 |
|          GetByteCount |    master |           16 |    580.4 ns |   6.738 ns |   6.302 ns |    580.1 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |           **32** |  **1,436.0 ns** |  **14.411 ns** |  **13.480 ns** |  **1,436.0 ns** |  **0.30** |    **0.01** |
|              GetBytes |    master |           32 |  4,737.9 ns | 107.426 ns | 100.486 ns |  4,746.6 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |           32 |    421.1 ns |   3.293 ns |   3.081 ns |    421.9 ns |  0.73 |    0.01 |
|          GetByteCount |    master |           32 |    578.3 ns |   8.748 ns |   7.755 ns |    577.6 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |           **64** |  **1,699.1 ns** |  **11.919 ns** |  **11.149 ns** |  **1,698.7 ns** |  **0.15** |    **0.00** |
|              GetBytes |    master |           64 | 11,014.0 ns | 109.592 ns | 102.513 ns | 10,997.3 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |           64 |    420.1 ns |   4.911 ns |   4.594 ns |    420.9 ns |  0.73 |    0.01 |
|          GetByteCount |    master |           64 |    578.5 ns |   3.912 ns |   3.468 ns |    579.5 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|              **GetBytes** |     **local** |          **128** |  **2,269.7 ns** |  **21.868 ns** |  **20.455 ns** |  **2,264.9 ns** |  **0.11** |    **0.00** |
|              GetBytes |    master |          128 | 19,900.3 ns | 247.904 ns | 231.890 ns | 19,861.0 ns |  1.00 |    0.00 |
|                       |           |              |             |            |            |             |       |         |
|          GetByteCount |     local |          128 |    423.0 ns |   5.200 ns |   4.610 ns |    422.1 ns |  0.73 |    0.01 |
|          GetByteCount |    master |          128 |    579.8 ns |   5.403 ns |   5.054 ns |    578.6 ns |  1.00 |    0.00 |

/cc @tannergooding and @ahsonkhan who I believe had some interest in taking advantage of these APIs within CoreLib.

__Not in this PR:__

- The refactoring to make the existing `ASCIIEncoding` class ride on top of this, as I didn't want to put too many things into this PR.
- The UTF-8 transcoding logic which rides on top of this as workhorse.