@GrabYourPitchforks 2019-02-11
Unit tests are at https://github.com/GrabYourPitchforks/corefx/blob/b83649c050964a3858c9dec2e48392afef83842e/src/System.Runtime/tests/System/Text/AsciiUtilityTests.netcoreapp.cs if you care to look at them. I don't know whether these unit tests belong in coreclr or corefx given the class itself is internal.

Given the complexity of these methods I believe it prudent to have unit tests which exercises them directly along with all of their edge cases and boundary conditions, _in addition to_ existing unit tests which exercise `ASCIIEncoding` and other public APIs.