@GrabYourPitchforks 2019-02-11
The experimental `ASCIIEncoding` replumbing is at https://github.com/GrabYourPitchforks/coreclr/blob/3de1f9d53e202278eb141403658aae21bac56d8c/src/System.Private.CoreLib/shared/System/Text/ASCIIEncoding.cs. In particular, search that file for the term _AsciiUtility_ and you'll see the workhorse methods being used as implementation.

/cc @layomia as FYI.