@jkotas 2019-02-11
> The experimental `ASCIIEncoding` replumbing

This replumbing should be included as part of this PR, I think.