@stephentoub 2019-02-11
> This replumbing should be included as part of this PR, I think.

Agreed. Otherwise this PR is adding dead, untested code that duplicates existing functionality.