@tannergooding 2019-02-11
> who I believe had some interest in taking advantage of these APIs within CoreLib.

@GrabYourPitchforks, I think one example would be: https://github.com/dotnet/coreclr/blob/70cf6d489032f42c9f643d59825c994c0402a3f7/src/System.Private.CoreLib/shared/System/Buffers/Text/Utf8Formatter/Utf8Formatter.Float.cs#L110-L114