@GrabYourPitchforks 2019-02-11
@jkotas Here are the numbers for replacing `ref byte` with `byte*` everywhere.

Calls to `Utf8Encoder.GetCharCount` for inputs of various length. In the below numbers, _Experiment-A_ is the PR as originally proposed (`ref byte` and friends), and _Experiment-B_ is using pointers (`byte*` everywhere).

|       Method |          Toolchain | DataLength |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------------- |------------------- |----------- |-----------:|----------:|----------:|------:|--------:|
| **GetCharCount** |       **Experiment-A** |          **0** |   **4.553 us** | **0.0904 us** | **0.2041 us** |  **0.30** |    **0.02** |
| GetCharCount |       Experiment-B |          0 |  12.437 us | 0.2841 us | 0.3040 us |  0.83 |    0.05 |
| GetCharCount | baseline (27411-2) |          0 |  15.185 us | 0.2981 us | 0.6022 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **1** |  **11.122 us** | **0.2208 us** | **0.4658 us** |  **0.70** |    **0.05** |
| GetCharCount |       Experiment-B |          1 |  19.321 us | 0.3851 us | 0.7602 us |  1.23 |    0.08 |
| GetCharCount | baseline (27411-2) |          1 |  15.820 us | 0.3127 us | 0.8664 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **2** |  **10.938 us** | **0.1867 us** | **0.2737 us** |  **0.64** |    **0.03** |
| GetCharCount |       Experiment-B |          2 |  19.517 us | 0.3830 us | 0.7003 us |  1.13 |    0.05 |
| GetCharCount | baseline (27411-2) |          2 |  17.215 us | 0.3443 us | 0.6208 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **3** |  **11.409 us** | **0.2215 us** | **0.2274 us** |  **0.65** |    **0.03** |
| GetCharCount |       Experiment-B |          3 |  19.766 us | 0.3869 us | 0.6248 us |  1.13 |    0.06 |
| GetCharCount | baseline (27411-2) |          3 |  17.633 us | 0.3485 us | 0.6797 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **4** |  **11.242 us** | **0.2234 us** | **0.4409 us** |  **0.61** |    **0.04** |
| GetCharCount |       Experiment-B |          4 |  19.762 us | 0.3916 us | 0.7545 us |  1.08 |    0.06 |
| GetCharCount | baseline (27411-2) |          4 |  18.399 us | 0.3667 us | 0.8426 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **8** |  **11.420 us** | **0.2279 us** | **0.4758 us** |  **0.51** |    **0.03** |
| GetCharCount |       Experiment-B |          8 |  19.194 us | 0.3839 us | 0.7397 us |  0.87 |    0.04 |
| GetCharCount | baseline (27411-2) |          8 |  22.235 us | 0.4414 us | 0.9213 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **12** |  **11.389 us** | **0.2267 us** | **0.4259 us** |  **0.45** |    **0.03** |
| GetCharCount |       Experiment-B |         12 |  19.651 us | 0.3924 us | 0.8362 us |  0.77 |    0.04 |
| GetCharCount | baseline (27411-2) |         12 |  25.734 us | 0.5044 us | 0.8145 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **16** |  **11.042 us** | **0.2201 us** | **0.3796 us** |  **0.37** |    **0.02** |
| GetCharCount |       Experiment-B |         16 |  18.870 us | 0.3647 us | 0.4612 us |  0.64 |    0.04 |
| GetCharCount | baseline (27411-2) |         16 |  29.419 us | 0.5872 us | 1.5470 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **32** |  **10.612 us** | **0.2120 us** | **0.2972 us** |  **0.22** |    **0.01** |
| GetCharCount |       Experiment-B |         32 |  18.993 us | 0.3785 us | 0.6728 us |  0.40 |    0.02 |
| GetCharCount | baseline (27411-2) |         32 |  47.434 us | 0.9461 us | 1.5808 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **64** |  **12.185 us** | **0.1919 us** | **0.1795 us** |  **0.19** |    **0.00** |
| GetCharCount |       Experiment-B |         64 |  19.541 us | 0.3815 us | 0.4541 us |  0.30 |    0.01 |
| GetCharCount | baseline (27411-2) |         64 |  64.689 us | 1.2710 us | 1.1889 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |        **128** |  **13.461 us** | **0.2671 us** | **0.3831 us** |  **0.13** |    **0.01** |
| GetCharCount |       Experiment-B |        128 |  21.831 us | 0.4334 us | 0.5481 us |  0.20 |    0.01 |
| GetCharCount | baseline (27411-2) |        128 | 107.397 us | 2.0725 us | 2.7667 us |  1.00 |    0.00 |