@jkotas 2019-02-11
`fixed` adds an overhead of a few instructions and a stack local. I can believe that this overhead would  show up for very small lengths. I have hard time believing that this overhead would cause this much difference for `DataLength` of 128.