@GrabYourPitchforks 2019-02-12
@jkotas Sure. Here are the relevant links:

https://github.com/GrabYourPitchforks/coreclr/blob/ascii_perf/src/System.Private.CoreLib/shared/System/Text/ASCIIEncoding.cs#L541 shows the implementation of `AsciiEncoding.GetCharCount`, using `AsciiUtility.GetIndexOfFirstNonAsciiByte` as a workhorse implementation.

https://github.com/GrabYourPitchforks/coreclr/commit/f579048803571d473b1e694daa64f04df408ab7e is the change that was made between _Experiment A_ and _Experiment B_. It was the simplest possible change I could think of to measure the cost of pinning the reference without touching any other logic.