@GrabYourPitchforks 2019-02-12
@jkotas Here's the run where the logic was rewritten to use `byte*` rather than `ref byte` everywhere. The code diff is at https://github.com/GrabYourPitchforks/coreclr/commit/cec62a029fe14de3972d5a858458e62757dfa779.

I was able to improve the performance slightly over the previous `byte*`-based iteration by tweaking a few code paths and controlling jumps more tightly. The throughput is still not quite as good as the `ref byte`-based code path, though.

|       Method |          Toolchain | DataLength |       Mean |     Error |    StdDev |    Median | Ratio | RatioSD |
|------------- |------------------- |----------- |-----------:|----------:|----------:|----------:|------:|--------:|
| **GetCharCount** |       **Experiment-A** |          **0** |   **4.270 us** | **0.0853 us** | **0.2461 us** |  **4.290 us** |  **0.29** |    **0.02** |
| GetCharCount |       Experiment-D |          0 |  12.230 us | 0.2408 us | 0.3046 us | 12.177 us |  0.84 |    0.04 |
| GetCharCount | baseline (27411-2) |          0 |  14.746 us | 0.2931 us | 0.5359 us | 14.740 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **1** |  **11.052 us** | **0.2178 us** | **0.4780 us** | **11.121 us** |  **0.68** |    **0.04** |
| GetCharCount |       Experiment-D |          1 |  16.601 us | 0.3297 us | 0.8743 us | 16.540 us |  1.02 |    0.07 |
| GetCharCount | baseline (27411-2) |          1 |  16.341 us | 0.3191 us | 0.8063 us | 16.495 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **2** |  **11.138 us** | **0.2064 us** | **0.1830 us** | **11.156 us** |  **0.67** |    **0.05** |
| GetCharCount |       Experiment-D |          2 |  17.591 us | 0.3503 us | 0.7540 us | 17.686 us |  1.04 |    0.07 |
| GetCharCount | baseline (27411-2) |          2 |  16.809 us | 0.3478 us | 1.0254 us | 16.800 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **3** |  **11.220 us** | **0.2243 us** | **0.4267 us** | **11.130 us** |  **0.66** |    **0.05** |
| GetCharCount |       Experiment-D |          3 |  18.322 us | 0.3573 us | 0.5562 us | 18.360 us |  1.07 |    0.06 |
| GetCharCount | baseline (27411-2) |          3 |  17.123 us | 0.3409 us | 0.9727 us | 17.178 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **5** |  **11.194 us** | **0.2214 us** | **0.4766 us** | **11.117 us** |  **0.59** |    **0.04** |
| GetCharCount |       Experiment-D |          5 |  17.843 us | 0.3560 us | 0.8665 us | 17.960 us |  0.93 |    0.06 |
| GetCharCount | baseline (27411-2) |          5 |  18.986 us | 0.3792 us | 0.7998 us | 19.065 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |          **7** |  **11.140 us** | **0.2159 us** | **0.3096 us** | **11.122 us** |  **0.55** |    **0.04** |
| GetCharCount |       Experiment-D |          7 |  17.833 us | 0.3542 us | 0.8210 us | 17.773 us |  0.88 |    0.06 |
| GetCharCount | baseline (27411-2) |          7 |  20.287 us | 0.4055 us | 1.0466 us | 20.452 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **13** |  **11.140 us** | **0.2222 us** | **0.4488 us** | **11.200 us** |  **0.44** |    **0.02** |
| GetCharCount |       Experiment-D |         13 |  17.875 us | 0.3568 us | 0.8340 us | 17.970 us |  0.71 |    0.05 |
| GetCharCount | baseline (27411-2) |         13 |  25.266 us | 0.5033 us | 1.0617 us | 25.247 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **16** |  **10.276 us** | **0.2019 us** | **0.4125 us** | **10.280 us** |  **0.37** |    **0.02** |
| GetCharCount |       Experiment-D |         16 |  16.288 us | 0.3255 us | 0.8106 us | 16.409 us |  0.59 |    0.04 |
| GetCharCount | baseline (27411-2) |         16 |  27.612 us | 0.5434 us | 1.2265 us | 27.738 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **32** |  **10.262 us** | **0.2035 us** | **0.4676 us** | **10.099 us** |  **0.24** |    **0.01** |
| GetCharCount |       Experiment-D |         32 |  16.455 us | 0.3286 us | 0.9637 us | 16.558 us |  0.38 |    0.03 |
| GetCharCount | baseline (27411-2) |         32 |  43.587 us | 0.8665 us | 2.0593 us | 43.141 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |         **64** |  **11.030 us** | **0.2184 us** | **0.3401 us** | **11.062 us** |  **0.18** |    **0.01** |
| GetCharCount |       Experiment-D |         64 |  17.394 us | 0.3414 us | 0.6975 us | 17.210 us |  0.28 |    0.02 |
| GetCharCount | baseline (27411-2) |         64 |  62.540 us | 1.2343 us | 2.4363 us | 61.952 us |  1.00 |    0.00 |
|              |                    |            |            |           |           |           |       |         |
| **GetCharCount** |       **Experiment-A** |        **128** |  **12.416 us** | **0.2473 us** | **0.4823 us** | **12.357 us** |  **0.12** |    **0.01** |
| GetCharCount |       Experiment-D |        128 |  19.770 us | 0.3948 us | 0.9383 us | 19.860 us |  0.20 |    0.01 |
| GetCharCount | baseline (27411-2) |        128 | 100.625 us | 2.0068 us | 3.6696 us | 98.576 us |  1.00 |    0.00 |